requirements.txt
# 记忆工程 v2.0 依赖
# ─── 人脸识别（新引擎） ─────────────────────────────────────────
insightface>=0.7.3
onnxruntime>=1.18
faiss-cpu>=1.7.4
opencv-python-headless>=4.9
mediapipe>=0.10.20

# ─── VLM/LLM API ────────────────────────────────────────────────
requests>=2.31.0
google-genai>=1.0.0

# ─── 图像处理 ───────────────────────────────────────────────────
pillow>=10.0.0
pillow-heif>=0.13.0
numpy>=1.24.0

# ─── EXIF / 地理 ────────────────────────────────────────────────
ExifRead>=3.0.0
geopy>=2.3.0

# ─── 通用 ────────────────────────────────────────────────────────
python-dotenv>=1.0.0
